==> code/analysis/14_prs_deg/14_p_value_histogram.R <==
# Load Libraries ----
suppressPackageStartupMessages({
  library(tidyverse)
  library(here)
  library(sessioninfo)
})

# Load Data ----
## Load PRS DEGs results ----
test_res <- read_csv(
  file = here(
    # "processed-data/rds/14_prs_deg",
    # "PRS_DEG_test_res_PRECAST07_donor_spd.csv"
    # "processed-data/rds/14_prs_deg",
    # "norm_PRS_deg/norm_PRS_DEG_test_res_PRECAST07_donor_spd.csv"

    "processed-data/rds/14_prs_deg/norm_PRS_DEG_test_res_PRECAST07_donor_spd.csv"
  )
)

# Plot histogram of p-values ----

hist(
  test_res$P.Value,
  breaks = 20, # bin width = 0.05
  main = "Histogram of p-value for PRS DEGs",
  xlab = "p-value",
  col = "lightblue",
  border = "black"
)
